refactor: one shared SlotConfig, plus a JS handle
`scalus.cardano.ledger.SlotConfig` had no shared definition: jvm/ and
native/ declared a byte-identical `Long` case class, and js/ a separate
`Double`, `js.Object`-extending, `@JSExportTopLevel("SlotConfig")` class.
Shared code - `CardanoInfo`, and `LedgerToPlutusTranslation`, which
computes the validity bounds a Plutus script sees - compiled against
whichever its platform supplied, and papered over the difference with
`.toLong` in fifteen places.

There is now one shared `Long` case class mixing in `SlotConfigPlatform`,
a platform trait carrying `slotToInstant`/`instantToSlot` on JVM and
Native and empty on JS. The class JavaScript sees is the new
`JsSlotConfig` handle, which keeps the exported name, the `number`
members, the constructor shape and the `Double` arithmetic the npm
package published - `timeToSlot` still returns a fractional slot, so
`scalus.d.ts` is unchanged.

`SlotConfigParityTest` still passes 7/7 on JVM, JS and Native.

Two behaviour notes: `JEmulator`'s `slotConfig == SlotConfig.mainnet`
becomes structural rather than reference equality, so a hand-built
mainnet config now also picks up mainnet protocol parameters; and
`JsCardanoInfo.slotConfig` returns a fresh handle per call instead of the
exported singleton, so JS identity comparison against `SlotConfig.mainnet`
no longer holds (field values are unchanged).

MiMa filters cover the two Scala.js artifacts, whose Scala signatures move
from `Double` to `Long`. Their JavaScript signatures do not change, and
the only way to call them is the npm bundle, which is relinked and
republished from this repo.

+++ scalus-testkit/shared/src/main/scala/scalus/testing/StandardTxVariations.scala
@@ -311,7 +311,7 @@ object StandardTxVariations {
             txTemplate: TxTemplate
         )(using ExecutionContext): Future[Seq[Transaction]] = {
             val env = reader.cardanoInfo
-            val currentSlot = env.slotConfig.timeToSlot(System.currentTimeMillis()).toLong
+            val currentSlot = env.slotConfig.timeToSlot(System.currentTimeMillis())
             val oneYearInSlots = 31536000L
 
             txTemplate.builder
